tools/requirements.txt
# Python requirements for ThermortoZigBee tools

# MQTT client for power monitoring
paho-mqtt>=1.6.1

# Data visualization
matplotlib>=3.7.1
numpy>=1.24.3

# Serial communication for debugging
pyserial>=3.5

# YAML parsing for Home Assistant configs
pyyaml>=6.0

# HTTP requests for OTA updates
requests>=2.31.0

# Data analysis
pandas>=2.0.3

# Progress bars
tqdm>=4.65.0